tools/requirements.txt
Pillow==8.1.0
pyserial==3.5
